refactor(components): Rename/create folders
