+++ app/(routes)/about-me/page.tsx
@@ -1,6 +1,6 @@
-import AvatarServices from "@/components/avatar-services";
-import CircleImage from "@/components/circle-image";
-import SliderServices from "@/components/slider-services";
+import AvatarServices from "@/components/routes/avatar-services";
+import CircleImage from "@/components/routes/circle-image";
+import SliderServices from "@/components/routes/slider-services";
 import Link from "next/link";
 // import TransitionPage from "@/components/transition-page";
