+++ components/routes/avatar-code.tsx
@@ -2,7 +2,7 @@
 
 import Image from 'next/image';
 
-import { MotionTransition } from './transition-component';
+import { MotionTransition } from '../main/transition-component';
 
 const AvatarCode = () => {
     return (
