+++ components/routes/avatar.tsx
@@ -3,7 +3,7 @@
 
 import Image from "next/image";
 
-import { MotionTransition } from "./transition-component";
+import { MotionTransition } from "../main/transition-component";
 
 export function Avatar() {
     return (
